requirements.txt
aiohttp==3.8.6
aiosqlite==0.19.0
loguru==0.7.2
quart
pyCryptoDome
bcrypt
PyExecJS